feat: Add static html output type.
