+++ packages/compiler/src/output/index.js
@@ -1,4 +1,5 @@
 export { default as ast } from './ast/index.js';
 export { default as html } from './html/index.js';
+export { default as static } from './html/static.js';
 export { default as latex } from './latex/index.js';
 export { default as esm } from './esm/index.js';
